feat(ingestion): staging-first paired-end upload, real-data harness
Two bodies of work land together because they touch the same backend
service layer and can't be split cleanly.

**Ingestion UX perfection pass**

- Staging-first picker: files accumulate across multiple picks into a
  frontend-only LaneStagingState slice. The upload session is only
  created once paired R1+R2 is satisfied, eliminating the parallel-
  session bug where picking R1 then R2 spawned two single-file sessions.
- inferReadPair.ts utility mirrors the backend READ_PAIR_PATTERN regex
  and normalize_fastq_sample_stem so the gate runs in the browser before
  any bytes leave the machine.
- LaneStagingPanel shows detected R1/R2 per file, sample stem, remove
  links, a live validation reason, and an outline Start upload button
  that stays disabled until validation.state === "ready".
- LaneReattachPanel replaces the awkward "Retry"/"Reselect" ghost buttons
  for the refresh-mid-upload case. Inline editorial panel with original
  filenames, per-file progress, fingerprint-matching drop strip, and a
  subordinate "discard session and start over" link.
- Commit-failure recovery gets its own affordance: "Try normalization
  again" outline button plus a "discard session" link. In-memory paused
  state shows "Resume upload · {remaining bytes}".
- Preview retry replaced with silent auto-retry once (tracked by
  autoRetryUsed) then an inline "sample again" editorial link.
- Backend validate_lane_files now rejects single-end at commit time with
  "Paired-end required" blocking issues. ReadLayout.SINGLE enum stays
  for legacy committed batches.
- New DELETE /api/workspaces/{id}/ingestion/sessions/{session_id}
  endpoint + delete_upload_session service fn that aborts in-flight
  multipart uploads and removes the session record.
- getCompactIssueLabel gains mappings for the new error strings.
- Playwright smoke test updated to click Start upload after staging;
  new "accumulates files across multiple picks" and "blocks single-end
  staging" cases. Two backend tests inverted to assert paired-end
  rejection; new test_delete_upload_session_clears_lane_state.
- docs/index.html landing mock updated to show the new staging panel
  (Normal lane) next to the ready state (Tumor lane), with updated
  section copy telegraphing the paired-end gate.

**Real-data ingestion harness**

- SEQC2 HCC1395 WES smoke fixtures via scripts/fetch_seqc2_sample_data.py
- .github/workflows/ci.yml + real-data-ingestion.yml
- Postgres + MinIO docker-compose infra, .run configs, .env.example
- backend/tests/test_real_data_ingestion.py
- playwright.config.ts + tests/e2e/ingestion-real-data.spec.ts
- Database refactor (backend/app/db.py, models/records.py, schemas.py)
  and s3_storage to back the resumable multipart upload path.

Verified: npx tsc --noEmit clean, eslint clean, pytest 19/19 passing.

+++ backend/app/models/records.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from typing import Optional
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String, Text, UniqueConstraint
+from sqlalchemy import BigInteger, DateTime, ForeignKey, Integer, String, Text, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db import Base
@@ -76,7 +76,7 @@ class WorkspaceFileRecord(Base):
     status: Mapped[str] = mapped_column(String(32), nullable=False)
     read_pair: Mapped[str] = mapped_column(String(32), nullable=False)
     storage_key: Mapped[str] = mapped_column(String(1024), nullable=False, unique=True)
-    size_bytes: Mapped[int] = mapped_column(Integer, nullable=False)
+    size_bytes: Mapped[int] = mapped_column(BigInteger, nullable=False)
     uploaded_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
     error: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
 
@@ -124,10 +124,10 @@ class UploadSessionFileRecord(Base):
     filename: Mapped[str] = mapped_column(String(512), nullable=False)
     format: Mapped[str] = mapped_column(String(32), nullable=False)
     read_pair: Mapped[str] = mapped_column(String(32), nullable=False)
-    size_bytes: Mapped[int] = mapped_column(Integer, nullable=False)
-    uploaded_bytes: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
+    size_bytes: Mapped[int] = mapped_column(BigInteger, nullable=False)
+    uploaded_bytes: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
     total_parts: Mapped[int] = mapped_column(Integer, nullable=False)
-    last_modified_ms: Mapped[int] = mapped_column(Integer, nullable=False)
+    last_modified_ms: Mapped[int] = mapped_column(BigInteger, nullable=False)
     fingerprint: Mapped[str] = mapped_column(String(1024), nullable=False, index=True)
     content_type: Mapped[Optional[str]] = mapped_column(String(255), nullable=True)
     storage_key: Mapped[str] = mapped_column(String(1024), nullable=False, unique=True)
@@ -156,7 +156,7 @@ class UploadSessionPartRecord(Base):
     )
     part_number: Mapped[int] = mapped_column(Integer, nullable=False)
     etag: Mapped[str] = mapped_column(String(255), nullable=False)
-    size_bytes: Mapped[int] = mapped_column(Integer, nullable=False)
+    size_bytes: Mapped[int] = mapped_column(BigInteger, nullable=False)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
 
     session_file: Mapped["UploadSessionFileRecord"] = relationship(back_populates="parts")
